feat: add play button ti demo
